mkswap: Set UUID for swap space
Sent to me by: Martin Schulze <joey@infodrom.org>:

A friend of mine is looking into the possibility of cloning Debian
(and other) systems automatically and stomped over swap partitions
getting assigned new UUIDs every time the new harddisk is partitioned
and swap is created.

His reseach lead him to

https://bugs.launchpad.net/ubuntu/+source/util-linux/+bug/66637

which refers to a patched version of mkswap:

http://free.nchc.org.tw/drbl-core/old/src/mkswap-uuid-0.1.0-2.tar.bz2
I'm attaching it for reference and in case the above URL becomes void.

It would be nice if you could apply this patch and send it upstream.

It's essential when partitions are to be recognised by their uuid and
not by their old device path anymore.

Regards,

        Joey

Addresses-Debian-Bug: 462662
Addresses-Ubuntu-Bug: 66637
Signed-off-by: LaMont Jones <lamont@debian.org>