hwclock: only call --systz from the udev rule
Use the new --systz option in the udev rule for /dev/rtc0, stepping the
clock if the hardware clock wasn't in localtime and updating the kernel
timezone.

Since udev can directly import /etc/default/rcS to get the UTC and
BADYEAR settings, we don't need a wrapper shell script for this saving a
bit of time.